docs: clean dev changelog history
